Corrigiendo algunas cosas

+++ SegundoParcial/Entidades/Articulos.cs
@@ -12,10 +12,10 @@ namespace SegundoParcial.Entidades
 
         public int ArticuloId { get; set; }
         public string Descripcion { get; set; }
-        public int Costo { get; set; }
-        public int Precio { get; set; }
-        public int Ganancia { get; set; }
-        public int Inventario { get; set; }
+        public double Costo { get; set; }
+        public double Precio { get; set; }
+        public double Ganancia { get; set; }
+        public double Inventario { get; set; }
 
         public Articulos()
         {
